# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopWave E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  stock: number;
  imageUrl: string;
  images?: string[];
  rating: number;
  reviewCount: number;
  featured: boolean;
  tags?: string[];
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  categoryId: number;
  stock: number;
  imageUrl: string;
  images?: string[];
  featured?: boolean;
  tags?: string[];
}

export interface ProductPatch {
  name?: string;
  description?: string;
  price?: number;
  /** @nullable */
  salePrice?: number | null;
  categoryId?: number;
  stock?: number;
  imageUrl?: string;
  images?: string[];
  featured?: boolean;
  tags?: string[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  imageUrl: string;
  productCount: number;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description: string;
  imageUrl: string;
}

export interface CartItem {
  id: number;
  productId: number;
  quantity: number;
  product: Product;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemPatch {
  quantity: number;
}

export interface WishlistItem {
  id: number;
  productId: number;
  product: Product;
  createdAt: string;
}

export interface WishlistInput {
  productId: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderItem {
  id: number;
  productId: number;
  quantity: number;
  price: number;
  productName: string;
  productImage: string;
}

export interface Order {
  id: number;
  items: OrderItem[];
  subtotal: number;
  shippingCost?: number;
  total: number;
  status: OrderStatus;
  shippingAddress: string;
  paymentMethod?: string;
  createdAt: string;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export interface OrderInput {
  shippingAddress: string;
  paymentMethod: string;
  items?: OrderItemInput[];
}

export type OrderStatusPatchStatus = typeof OrderStatusPatchStatus[keyof typeof OrderStatusPatchStatus];


export const OrderStatusPatchStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusPatch {
  status: OrderStatusPatchStatus;
}

export interface Review {
  id: number;
  productId: number;
  rating: number;
  comment: string;
  authorName: string;
  createdAt: string;
}

export interface ReviewInput {
  productId: number;
  rating: number;
  comment: string;
  authorName: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface DashboardSummary {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  totalCategories: number;
  recentOrders: Order[];
  topProducts: Product[];
  ordersByStatus: StatusCount[];
}

export type ListProductsParams = {
search?: string;
categoryId?: number;
minPrice?: number;
maxPrice?: number;
sort?: ListProductsSort;
featured?: boolean;
onSale?: boolean;
limit?: number;
offset?: number;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  newest: 'newest',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  rating: 'rating',
  popular: 'popular',
} as const;

export type ListReviewsParams = {
productId: number;
};