# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ShopWave E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart items
      responses:
        "200":
          description: Cart items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItem"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Cart item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear entire cart
      responses:
        "204":
          description: Cart cleared
  /cart/{id}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemPatch"
      responses:
        "200":
          description: Updated cart item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusPatch"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, description, price, categoryId, stock, imageUrl, rating, reviewCount, featured, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string }
        description: { type: string }
        price:       { type: number }
        salePrice:   { type: ["number", "null"] }
        categoryId:  { type: integer }
        categoryName: { type: ["string", "null"] }
        stock:       { type: integer }
        imageUrl:    { type: string }
        images:      { type: array, items: { type: string } }
        rating:      { type: number }
        reviewCount: { type: integer }
        featured:    { type: boolean }
        tags:        { type: array, items: { type: string } }
        createdAt:   { type: string }
...
    Category:
      type: object
      required: [id, name, slug, description, imageUrl, productCount]
      properties:
        id:           { type: integer }
        name:         { type: string }
        slug:         { type: string }
        description:  { type: string }
        imageUrl:     { type: string }
        productCount: { type: integer }
    CategoryInput:
      type: object
      required: [name, slug, description, imageUrl]
      properties:
        name:        { type: string, minLength: 1 }
        slug:        { type: string, minLength: 1 }
        description: { type: string }
        imageUrl:    { type: string }
    CartItem:
      type: object
      required: [id, productId, quantity, product]
      properties:
        id:        { type: integer }
        productId: { type: integer }
        quantity:  { type: integer }
        product:
          $ref: "#/components/schemas/Product"
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity:  { type: integer }
    CartItemPatch:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer }
...
    Order:
      type: object
      required: [id, items, subtotal, total, status, shippingAddress, createdAt]
      properties:
        id:              { type: integer }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:        { type: number }
        shippingCost:    { type: number }
        total:           { type: number }
        status:          { type: string, enum: [pending, processing, shipped, delivered, cancelled] }
        shippingAddress: { type: string }
        paymentMethod:   { type: string }
        createdAt:       { type: string }
    OrderItem:
      type: object
      required: [id, productId, quantity, price, productName, productImage]
      properties:
        id:           { type: integer }
        productId:    { type: integer }
        quantity:     { type: integer }
        price:        { type: number }
        productName:  { type: string }
        productImage: { type: string }
    OrderInput:
      type: object
      required: [shippingAddress, paymentMethod]
      properties:
        shippingAddress: { type: string }
        paymentMethod:   { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity:  { type: integer }
    OrderStatusPatch:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [pending, processing, shipped, delivered, cancelled] }
...