# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopWave E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  items: OrderItem[];
  subtotal: number;
  shippingCost?: number;
  total: number;
  status: OrderStatus;
  shippingAddress: string;
  paymentMethod?: string;
  createdAt: string;
}