# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopWave E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  shippingAddress: string;
  paymentMethod: string;
  items?: OrderItemInput[];
}