# lib/api-zod/src/generated/types/orderStatusPatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopWave E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatusPatchStatus } from './orderStatusPatchStatus';

export interface OrderStatusPatch {
  status: OrderStatusPatchStatus;
}