# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopWave E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  stock: number;
  imageUrl: string;
  images?: string[];
  rating: number;
  reviewCount: number;
  featured: boolean;
  tags?: string[];
  createdAt: string;
}